added search bar to user menu, failed to make friends work

+++ src/main/resources/db/migration/V2__init.sql
@@ -3,6 +3,6 @@ DROP TABLE IF EXISTS posts;
 CREATE TABLE posts (
   id bigserial PRIMARY KEY,
   content varchar(250) NOT NULL,
-    user_id INT,
+    user_id BIGINT,
     FOREIGN KEY(user_id) REFERENCES users(id)
 );
